Modify the startup sequence
The Java VM sets up its own signal handers, so the setup of our own
handlers must happen after the plugin initialization, and the installed
handlers should do proper chaining.

The Java VM also does not like the daemon() call after it has started.
It gets confused about which is the main thread and it cannot be shut
down properly. So move the daemon() call after the plugin initialization
as well.

git-svn-id: https://edges-3g-bridge.svn.sourceforge.net/svnroot/edges-3g-bridge/apps/3g-bridge/trunk@2302 a7169a2c-3604-0410-bc95-c702d8d87f7a