Fixed bugzilla bug #181.
If the DLMgr is not configured, a job could stick in INIT. The QueueManager
would try to request downloading its input files over and over again.

Solution: `handled' flag in eventdata; the event handler should set this
flag. If the QueueManager finds it unset after raising the event, the job's
status is set to ERROR (because it will never be executed).

git-svn-id: https://edges-3g-bridge.svn.sourceforge.net/svnroot/edges-3g-bridge/3g-bridge/trunk@2665 a7169a2c-3604-0410-bc95-c702d8d87f7a