Removed logmon from WSSubmitter.
Thw WSSubmitter will not generate job_entry messages anymore.
Instead, the message will be generated by the Bridge, when the
plugin handles the job.

This delays the appearance of the message by a few seconds, but
the timestamp will still be accurate.

This enables us to reimplement the WSSubmitter in CGI/PHP, and thus,
to support authentication.

git-svn-id: https://edges-3g-bridge.svn.sourceforge.net/svnroot/edges-3g-bridge/3g-bridge/trunk@2719 a7169a2c-3604-0410-bc95-c702d8d87f7a